chore: refactored the folder structure and updated with latest main changes
